add crontab

+++ web/database/migrations/1654184420474_job_security_audits.ts
@@ -11,8 +11,7 @@ export default class JobSecurityAudits extends BaseSchema {
       table.integer('progress', 3).defaultTo(0)
       table.boolean('in_use').defaultTo(false)
       table.enum('status', [ 'waiting', 'running', 'completed', 'failed' ]).notNullable().defaultTo('waiting')
-      table.integer('vulns').defaultTo(0)
-      
+      table.double('vulns').defaultTo(0)
       table.string('pdf')
       table.timestamp('created_at', { useTz: true })
       table.timestamp('updated_at', { useTz: true })
